feat(plugin-layout): add setting for native app redirect
