+++ packages/apps/plugins/plugin-layout/src/components/LayoutSettings.tsx
@@ -14,8 +14,16 @@ export const LayoutSettings = ({ settings }: { settings: LayoutSettingsProps }) 
   const { t } = useTranslation(LAYOUT_PLUGIN);
 
   return (
-    <SettingsValue label={t('settings show footer label')}>
-      <Input.Switch checked={settings.showFooter} onCheckedChange={(checked) => (settings.showFooter = !!checked)} />
-    </SettingsValue>
+    <>
+      <SettingsValue label={t('settings show footer label')}>
+        <Input.Switch checked={settings.showFooter} onCheckedChange={(checked) => (settings.showFooter = !!checked)} />
+      </SettingsValue>
+      <SettingsValue label={t('settings native redirect label')}>
+        <Input.Switch
+          checked={settings.enableNativeRedirect}
+          onCheckedChange={(checked) => (settings.enableNativeRedirect = !!checked)}
+        />
+      </SettingsValue>
+    </>
   );
 };
