+++ packages/apps/plugins/plugin-layout/src/helpers.ts
@@ -9,3 +9,21 @@ export const uriToActive = (uri: string) => {
 
 export const activeToUri = (active?: string) =>
   '/' + (active ? active.split(':').map(encodeURIComponent).join('/') : '');
+
+// TODO(mjamesderocher): Factor out as part of NavigationPlugin.
+export const checkAppScheme = (url: string) => {
+  const iframe = document.createElement('iframe');
+  iframe.style.display = 'none';
+  document.body.appendChild(iframe);
+
+  iframe.src = url + window.location.pathname.replace(/^\/+/, '');
+
+  const timer = setTimeout(() => {
+    document.body.removeChild(iframe);
+  }, 3000);
+
+  window.addEventListener('pagehide', (event) => {
+    clearTimeout(timer);
+    document.body.removeChild(iframe);
+  });
+};
